requirements.txt
amqp==1.4.7
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.21
celery==3.1.19
cloudinary==1.1.3
coverage==3.7.1
coveralls==1.1
decorator==4.0.9
dj-database-url==0.3.0
Django==1.8.3
django-all-access==0.7.2
django-bootstrap-form==3.2
django-bower==5.0.4
django-celery==3.1.17
django-crispy-forms==1.6.0
django-endless-pagination==2.0
django-envie==0.0.9
django-filter==0.12.0
django-multiselectfield==0.1.3
django-postgrespool==0.3.0
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
docopt==0.6.2
funcsigs==0.4
gnureadline==6.3.3
gunicorn==19.3.0
hashids==1.1.0
httpie==0.9.3
ipdb==0.9.0
ipython==4.1.1
ipython-genutils==0.1.0
kombu==3.0.30
meld3==1.0.2
mock==1.0.1
oauthlib==1.0.3
optional-django==0.3.0
path.py==8.1.2
pbr==1.5.0
pexpect==4.0.1
pickleshare==0.6
Pillow==2.9.0
postgres==2.2.1
psycopg2==2.6.1
ptyprocess==0.5.1
pycrypto==2.6.1
pycurl==7.43.0
pyfirebase==1.3
Pygments==2.1.1
PyJWT==1.4.0
python-coveralls==2.4.3
python-openid==2.2.5
python-social-auth==0.2.19
pytz==2015.7
PyYAML==3.11
reportlab==3.2.0
requests==2.10.0
requests-mock==0.7.0
requests-oauthlib==0.6.1
sass==2.3
selenium==2.47.1
sendgrid==1.5.14
sh==1.11
simplegeneric==0.8.1
six==1.10.0
smtpapi==0.2.0
SQLAlchemy==1.0.8
sseclient==0.0.12
traitlets==4.1.0
webpack==6.0.0
whitenoise==2.0.3
